custom exception added for all apis

+++ src/main/java/com/livares/intern/service/impl/CartServiceImplementation.java
@@ -3,8 +3,6 @@ package com.livares.intern.service.impl;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import com.livares.intern.DTO.CartDTO;
@@ -16,7 +14,6 @@ import com.livares.intern.exception.ErrorCodes;
 import com.livares.intern.repository.ProductRepository;
 import com.livares.intern.repository.UserProductCartRepository;
 import com.livares.intern.repository.UsersRepository;
-import com.livares.intern.response.ResponseHandler;
 import com.livares.intern.service.CartService;
 
 @Service
@@ -31,8 +28,11 @@ public class CartServiceImplementation implements CartService {
 	@Override
 	public String addToCart(long userId, long productId) {
 
-		Users users = usersRepository.findById(userId).get();
-		Product product = productRepository.findById(productId).get();
+		Users users = usersRepository.findById(userId).orElse(null);
+		Product product = productRepository.findById(productId).orElse(null);
+		if (users == null || product == null) {
+			throw new CustomException(ErrorCodes.NOT_FOUND, "User or product is not found");
+		}
 		int count = userProductCartRepository.countOfProducts(userId, productId);
 		if (count == 0) {
 
@@ -41,7 +41,6 @@ public class CartServiceImplementation implements CartService {
 			userProductCart.setProductId(product);
 			userProductCartRepository.save(userProductCart);
 			return "Product added successfully";
-					
 
 		} else {
 			throw new CustomException(ErrorCodes.METHOD_NOT_SUPPORTED, "Product already added to cart");
@@ -52,22 +51,27 @@ public class CartServiceImplementation implements CartService {
 	@Override
 	public List<CartDTO> fetchDetails(long userId) {
 
-		return userProductCartRepository.fetchData(userId);
-				
+		List<CartDTO> cartDTOs = userProductCartRepository.fetchData(userId);
+		if (cartDTOs == null || cartDTOs.isEmpty()) {
+
+			throw new CustomException(ErrorCodes.NOT_FOUND, "Items not found with this id");
+		} else {
+			return cartDTOs;
+		}
 
 	}
 
 	@Override
 	public String deleteUser(long userId, long productId) {
-		try {
 
-			long id = userProductCartRepository.getCartById(userId, productId);
-			userProductCartRepository.deleteById(id);
-			return "deleted successfully";
-					
-		} catch (Exception e) {
-			throw new CustomException("user or Item is not found!");
+		UserProductCart cart = userProductCartRepository.findByUserId(userId, productId);
+		if (cart == null) {
+			throw new CustomException(ErrorCodes.NOT_FOUND, "Item not found in the cart with this id");
 		}
+		long id = cart.getId();
+
+		userProductCartRepository.deleteById(id);
+		return "deleted successfully";
 
 	}
